extracción de datos de ventas y clasificación según contado o crédito, y modificaciones al modelo de datos, seeder, y controlador

+++ database/migrations/2020_04_20_201232_create_productos_table.php
@@ -17,7 +17,7 @@ class CreateProductosTable extends Migration
             $table->id();
             $table->string('nombre');
             $table->unsignedInteger('cantidad');
-            $table->float('precio');
+            $table->decimal('precio', 13, 2);
             $table->softDeletes();
             $table->timestamps();
         });
